working on making adding courses show up immediatly instead of after refresh
