+++ frontend/src/components/CourseTable.vue
@@ -21,6 +21,7 @@
 import Semester from "./Semester.vue";
 import coursesJson from "../data/courses.json";
 import CourseSearch from "./CourseSearch.vue";
+import { useStore } from "../stores/store";
 
 export default {
     components: {
@@ -40,10 +41,15 @@ export default {
             coursesData: coursesJson,
             showSearch: false,
             opened: null,
+            store: useStore()
         };
     },
     mounted() {
         console.log(this.$refs);
+        this.store.$subscribe(() => {
+            console.log("something changed in coursetable");
+            // this.parsedSemesters = 
+        });
     },
     methods: {
         showAddCourseModal(semesterName) {
